Add a fold-add-into-dest pattern and pass (#966)
Fold a linalg.add into the dest/out of the contraction op of one of
its operands. The sum with the `other` operand will be performed by
having the contraction accumulate on top of `other`. Note that this
changes the result of the contraction - hence the linalg.add must be
its single use.

Main value is likely to come from that a buffer allocation can be 
elided, in case the contraction op needed a buffer to be allocated 
for its result.