Split reduction dim pass (#975)
Adds a pass to allow tiling contraction's innermost reduction dimension
using serial loop with in-place accumulation.

Compared to other available transformations, this rewrite computes
reduction sequentially with in-place accumulation which avoids temporary
allocation and separate reduction operation. This tiling strategy is
more friendly in terms of register and memory pressure more suitable for
low-level GPU kernel generation. Similarly, restriction to the innermost
dimension is there to simplify both usage and pass logic as the rewrite
is geared toward progressive GEMM lowering.

Additionally, a GPU vectorization control flag is added to allow
grouping of passes which target lowering through vector operations and
might not be compatible with other existing lowering strategies.
Effectively, this pass perform separate K-dim split which is currently
baked into linalg-to-xegpu lowering, and the two are not fully
compatible.