Lower vector.contract to chain of vector.fma for matrix multiply.
Implements the lowering of vector contraction op for GEMM of size MxN
to sequence of vector FMAs wrapped inside scf.for loop with iterargs
to accumulate the result of FMAs.The K dimension size is expected to
be 1. The idea is to exploit the AVX-512 feature to generate optimal
vector code.