Move `linalg-to-xsmm` before `linalg-to-vector` and add option to skip lowerings (#981)
In order to proceed with separate vectorization paths, we need to
improve the pipeline so that we can skip certain patterns to XSMM
depending on the path we take.

Before this PR, the path is:
```
bufferization -> <vectorize?> -Y-> linalg-to-vector ... -> lower to LLVM
                             \-N-> linalg-to-xsmm ----------^
```

After this patch, i becomes:
```
bufferization -> linalg-to-xsmm (partial) -> <vectorize?> -Y-> linalg-to-vector ... -> lower to LLVM
                                                         \-N-> -------------------------^
```

The `partial` part is controlled by a new option (`skip-operations`),
which can be passed via command line or changed in the pipeline builder
function from other command line options (ex. `vector-to-function`,
`vector-to-kernel`, etc.).

The state would end up as:
```
bufferization -> linalg-to-xsmm (partial) -> <vector-to-function?> -> linalg-to-vector -> vector-to-xsmm -> lower to LLVM
                                          | <vector-to-kernel> -> linalg-to-vector -> vector-to-kernel ------^
                                          \ <else> ----------------------------------------------------------^
```

Patterns that do not lower to neither XSMM, function or kernel will be
lowered to loop by the final cleanups.

@KavithaTipturMadhu, this should allow you to add a new flag (ex.
`vector-to-function`) where you enable only one operation at a time to
lower from vector to function and disable that in `linalg-to-xsmm` at
the same time. This helps #980.

@shahidact and @arun-thmn, similarly for #977 and #978, this should
allow you two to have a shared pipeline that goes from linalg to vector
all the way down to kernel but do so only for a single operation at a
time, which can be _different_ than Kavitha's order, controlled by
different lists from your different option.